// src/enums/transactions/transaction-statuses-enum.ts
export enum TransactionStatusesEnum {
    NEW = 'new',
    PENDING = 'pending',
    IN_PROCESS = 'in_process',
    INSUFFICIENT_FUNDS = 'insufficient_funds',
    TRANSFER_INITIATED = 'transfer_initiated',
    COMPLETED = 'completed',
    CANCELLED = 'cancelled',
    ERROR = 'error',
    REFUNDED = 'refunded',
}